move ConfigSelectingClientCall to inner class
